requirements.txt
Django
djangorestframework
dj-rest-auth
django-cors-headers
psycopg2-binary
pandas
scikit-learn
joblib
gunicorn